worker/pyproject.toml
[project]
name = "transcribee-worker"
version = "0.1.0"
description = "a transcription worker for the transcribee open-source transcription platform"
authors = [
    { name = "anuejn", email = "jarohabiger@googlemail.com" },
    { name = "pajowu", email = "git@ca.pajowu.de" },
    { name = "phlmn", email = "info@philipp-mandler.com" },
    { name = "vup", email = "robin.ole.heinemann+transcribee@gmail.com" },
]

dependencies = [
    "numpy>=1.23.5",
    "pydantic[dotenv]>=1.10.7",
    "transformers>=4.26.1",
    "torchaudio>=2.0.0",
    "torch>=2.0.0",
    "automerge @ git+https://github.com/transcribee/automerge-py.git@057303bd087401f12b166e2adb7161f0fcb3a9dc",
    "websockets>=10.4",
    "scikit-learn>=1.2.2",
    "watchfiles>=0.19.0",
    "speechbrain>=0.5.14",
    "ffmpeg-python>=0.2.0",
    "transcribee-proto @ file:///${PROJECT_ROOT}/../proto",
    "PyICU>=2.11",
    "faster-whisper>=0.10.0",
]

requires-python = ">=3.10"
readme = "README.md"
license = { text = "AGPL-3.0" }

[tool.pdm.dev-dependencies]
dev = ["black>=23.1.0", "inquirer>=3.1.3", "pyright>=1.1.314", "pytest>=7.3.1"]

[tool.pdm.scripts]
start = "python run.py"
test = "pytest tests/"

[build-system]
requires = ["pdm-pep517>=1.0"]
build-backend = "pdm.pep517.api"

[tool.pdm.resolution.overrides]
soundfile = "0.11.0"




[[tool.pdm.source]]
type = "find_links"
url = "https://download.pytorch.org/whl/cpu/torch_stable.html"
name = "torch"
